+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project roughly adheres to [Semantic Versioning](https://semver.org/spe
 - **Breaking:** `LLMLayerActivationsOp` has an `apply_chat_template=True` parameter, which allows you to use a raw text column as input. You will unfortunately have to recompute all embeddings.
 - **Breaking** `JinjaTemplatizeOp` now supports string and float constants in templates. You will unfortunately have to recompute all ops that depend on your templates.
 - Arrow writing operations now used fixed-size row groups and gentle zstd compression.
+- `krnel.graph.Runner()` now returns a singleton instance. This ensures that calling `kg.runner.Runner()` from multiple modules doesn't create a graph derived from multiple runner instances.
 
 ### Added
 - **New type:** Add a `JSONColumnType` using Arrow's support for structured data. This is ideal for multi-turn conversations etc.